Move runtime to use only Prost codec and Bytes to encode data passed around.
This change focuses on switching from using vector of bytes to using Bytes
crate that provides abstractions for working with bytes. It enables
efficient container for storing and operating on contiguous slices of memory.

Currently Raft implementation still uses in some places vectors of bytes. We
are going to address that separately in the Raft repository.

Change-Id: I939d3542e47b1feccea913c5ae89952467460c5c