Initial implementation of LedgerActor
This includes more or less complete implementation of LedgerActor that delegates most of the work to
the existing LedgerService, but doesn't do much on top of that yet. There is no support for snapshots
yet.

There is no additional test coverage either on top of unit tests that already exist in ledger.rs.
The plan is to work on the integration test next.

Change-Id: I14c9a04d9b4cd97e4aa66da39fee390c37342cc9