Add CommunicationModule trait and some default skeleton implementation.
This trait is responsible for managing communication between replicas
i.e. establishing a handshake if talking to a peer replica for the first
time, encrypting/decrypting messages to and from replicas, verifying
attestation and generating attestation reports that will be sent over to
the peer as part of handshake.

For now, the Encryptor and Attestor traits have been ommitted. They will
be added eventually once the basic handshake protocol structure is
implemented.

Change-Id: I8882dc5979b72a6298e9d87a0b2805a7abf246b5