Introducing application level message delivery mechanism to enable application to application and appliction to untrusted launcher communication.
The appliction message has the main unencrypted contents that is used by the untrusted
launcher to handle messages. The application message payload contains data that is
not meant to be interpretted by the untrusted launcher but rather forwarded to another
application or system for processing.

Note that each untrusted launcher and application how to deserialize and process the message
contents.

Change-Id: I541a734eab0cdc1c45c3072f1867cd7265bac3b4