Create skeleton of the Tablet Cache
Introduce Transaction Manager and Key Value Store abstractions,
wire them together through the Tablet Cache actor.

Transaction Manager is responsible for providing efficient cache
for the tablet metadata and contents, transactional support for
tablet processing.

Key Value Store is responsible for processing consumer requests
in terms of tablet trasactions.

Note we discussed offline this is work in progress and is expected
to evolve. The next step is produce Key Value Store implementation
against Transaction Manager interface and adjust it as needed. The
last step will be to implement the actual Transaction Manager.

Change-Id: Ifd7ed1d1707a5ef7450343b06d32b949a328ad07