Initial implementation of the Tablet Metadata Cache
Tablet Metadata Cache splits the consistent hashing ring
for each table into regions of configured size. When tablets
are requested to be resolved it checks if the region has
fresh data or stale. If the data is stale it requests
Tablet Store to list tablets that are responsible
for that region. If any tablet metadata is not up to
date the whole region is invalidated and requested again.

The change contains smoke test, more test converage will
be added in the subsequent changes.

Change-Id: I2be094852cc2d6c2ca1de5c14362a3b8a03301fd