Switch Tablet Store list op from tablet to key hash ranges
Initial implementation was using tablet id range in the list
op to find tablets. However with this api it is hard to implement
metadata cache in Tablet Cache. Instead for Tablet Cache it
is more convenient to ask Tablet Store to list all tablets that
are responsible for a key range. The differnce with tablet id
range is that the tablets that are responsile for the requested
key range may be outside of the key range. More specifically
must be searched for within the range and then clockwise from
the end of the range.

This change makes that switch. Separate change will introduce
support for multiple tables.

Change-Id: I4de67cd79689da97c4bd570d6e9738cbab25f52e