Introduce handling of Tablet Store ops execution errors
Errors from Tablet Store to process tablets op may not produce
any results but an error code. To simplify dispatch and processing
of such errors when corresponding request is created we
also create artificial Tablet Store responses marking all ops
and the request itself as failed. Then if an actual error is
passed we replace it with arfiticial response otherwise we
pass actual response.

Change-Id: I95543fcf48e548e94523f43bb0eb1d538a56af87