Transition to FAILED state if handshake message is received again in Completed state.
Receiving a handshake message in Completed state indicates that
something went wrong on the other side in the last attempt due to which
handshake is being retried. In this case, the state machine should
simply transition to Failed state so that handshake can be retried after
enough ticks have passed.

Change-Id: I9e3db92efb3186b03845ac1f4bfc1fd796eb524c