Add a timeout for a handshake in Initiated state.
Once enough ticks have passed in Initiated state without a response, the
state machine will transition to FAILED state and handshake will be
retried later. For now we set the default timeout to 3*heartbeat_tick to
provide enough time for a response to arrive. We can later expose this
setting as part of the RaftConfig if further tuning is needed.

Change-Id: I3283dac8f3448f594c19c43d3e1c574e61f6058f