Add a hook to obtain the ReferenceValues from the Actor instead of passing it down from the untrusted code.
The ReferenceValues are hard-coded and will be provided from within the
trusted code itself so that they cannot be tampered with. For now, the
ReferenceValues are populated with "skip" but they will eventually
contain endorsement based values which should not change between binary
versions. In subsequent changes, these ReferenceValues will be passed
down when creating a new Oak session.

Also, update hte Oak version to pull in the latest changes and fix some
breaks.

Change-Id: I3dc705fea279c52665ef998ab2d304caffdd2c7f