Add tcp_runtime support for Oak Containers.
In the same way that ApplicationService allows an Actor to be bound to a
micro_rpc server, TonicApplicationService allows an Actor to be bound to
a Tonic (gRPC) server. This struct is only available when building in
std environments.

Similarly, tcp_proto now includes Tonic server classes for
EndpointService; these classes are also only available in std
environments.

The Oak dependency has been updated to include
https://github.com/project-oak/oak/commit/1d022f22cf7817eb99ba050a7ae3b16bfa349bac.

Bug: 392130200
Change-Id: If2ee6f63c807bb442942eea06f79d794e71fd0e9