Use DefaultBindingKeyExtractor in TonicApplicationService.
When updating to a newer Oak revision, 7254e8c switched
TonicApplicationService from `SignatureBinderBuilder` + `InstanceSigner`
(which had been deleted) to `InstanceSessionBinder`; this caused the
session to switch from using the application signing key to the session
binding key. The peer verifier wasn't updated, causing errors when
establishing sessions. The fix is to switch from
`DefaultSigningKeyExtractor` (the default) to
`DefaultBindingKeyExtractor`.

`ApplicationService` is not updated since the signing key is still used
for Restricted Kernel sessions.

The `KeyExtractor` must now be configurable; it's passed in as an Arc to
avoid needing to create a new type of factory.

Change-Id: If965829f763b5e41e81b5254df31ed9b54da570c